refactor: Move `getTimezoneOffset` and `getLocalTime` to `locale.ts`
